Handle epoch datetimes as typed integers

+++ tests/integration/cases/scalar_datetime/Python_type_hints_always_dt_epoch.py
@@ -1,1 +1,1 @@
-my_data: float = 1705321800.0
+my_data: int = 1705321800
